+++ Rest/src/main/java/sber/spring/Rest/controllers/BinController.java
@@ -4,7 +4,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import sber.spring.Rest.service.BinService;
 import sber.spring.Rest.service.ClientService;
+import sber.spring.Rest.service.ProductService;
 
 import java.net.URI;
 import java.net.URISyntaxException;
@@ -13,44 +15,46 @@ import java.net.URISyntaxException;
 @RequestMapping("bin")
 public class BinController {
     @Autowired
-    private ClientService clientService;
+    private BinService binService;
+    @Autowired
+    private ProductService productService;
 
     @PutMapping("/{binId}/add/{productId}")
     public ResponseEntity<String> binAdd(@PathVariable int binId, @PathVariable int productId) throws URISyntaxException {
-        if(!clientService.searchProductRep(productId).isPresent()){
+        if(!productService.searchProductRep(productId).isPresent()){
             return new ResponseEntity<String>("Product with id = " + productId + " doesn't exists", HttpStatus.NOT_FOUND);
-        } else if(!clientService.searchBinRep(binId).isPresent()){
+        } else if(!binService.searchBinRep(binId).isPresent()){
             return new ResponseEntity<String>("bin with id = " + binId + " doesn't exists", HttpStatus.NOT_FOUND);
-        } else if (clientService.isInBin(binId, productId)){
+        } else if (binService.isInBin(binId, productId)){
             return new ResponseEntity<String>("You already added this product", HttpStatus.NOT_FOUND);
         } else {
-            int savedId = clientService.addToBin(binId, productId);
+            binService.addToBin(binId, productId);
             return new ResponseEntity<String>("Product added", HttpStatus.CREATED);
         }
     }
 
     @PutMapping("{binId}/change/{productId}/{quantity}")
     public ResponseEntity<String> binChangeQuantity(@PathVariable int binId, @PathVariable int productId, @PathVariable int quantity){
-        if(!clientService.searchProductRep(productId).isPresent()){
+        if(!productService.searchProductRep(productId).isPresent()){
             return new ResponseEntity<String>("Product with id = " + productId + " doesn't exists", HttpStatus.NOT_FOUND);
-        } else if(!clientService.searchBinRep(binId).isPresent()){
+        } else if(!binService.searchBinRep(binId).isPresent()){
             return new ResponseEntity<String>("bin with id = " + binId + " doesn't exists", HttpStatus.NOT_FOUND);
-        } else if (!clientService.isInBin(binId, productId)) {
+        } else if (!binService.isInBin(binId, productId)) {
             return new ResponseEntity<String>("You didn't add this product", HttpStatus.NOT_FOUND);
         } else {
-            clientService.changeQuantity(binId, productId, quantity);
+            binService.changeQuantity(binId, productId, quantity);
             return new ResponseEntity<String>("Quantity changed", HttpStatus.OK);
         }
     }
 
     @DeleteMapping("/{binId}/change/{productId}")
     public ResponseEntity<String> deleteFromBin(@PathVariable int binId, @PathVariable int productId){
-        if(!clientService.searchProductRep(productId).isPresent()){
+        if(!productService.searchProductRep(productId).isPresent()){
             return new ResponseEntity<String>("Product with id = " + productId + " doesn't exists", HttpStatus.NOT_FOUND);
-        } else if(!clientService.searchBinRep(binId).isPresent()){
+        } else if(!binService.searchBinRep(binId).isPresent()){
             return new ResponseEntity<String>("bin with id = " + binId + " doesn't exists", HttpStatus.NOT_FOUND);
         } else {
-            return clientService.deleteProductFromBin(binId, productId)
+            return binService.deleteProductFromBin(binId, productId)
                     ? new ResponseEntity<String>("Product deleted successfully", HttpStatus.NO_CONTENT)
                     : new ResponseEntity<String>("You didn't add this product", HttpStatus.NOT_FOUND);
         }
@@ -58,10 +62,10 @@ public class BinController {
 
     @PutMapping("/{binId}/payment")
     public ResponseEntity<Void> payment(@PathVariable int binId){
-        if (clientService.searchBinRep(binId).isPresent()) {
+        if (binService.searchBinRep(binId).isPresent()) {
             return ResponseEntity.notFound().build();
         } else {
-            return clientService.pay(binId)
+            return binService.pay(binId)
                     ? ResponseEntity.noContent().build()
                     : ResponseEntity.notFound().build();
         }
