+++ Rest/src/main/java/sber/spring/Rest/controllers/ClientController.java
@@ -4,7 +4,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import sber.spring.Rest.entities.Bin;
 import sber.spring.Rest.entities.Client;
+import sber.spring.Rest.service.BinService;
 import sber.spring.Rest.service.ClientService;
 
 import java.net.URISyntaxException;
@@ -16,9 +18,12 @@ public class ClientController {
     @Autowired
     private ClientService clientService;
 
+    @Autowired
+    private BinService binService;
+
     @PostMapping
     public ResponseEntity<String> clientReg(@RequestBody Client client) throws URISyntaxException {
-        long savedId = clientService.saveClient(client);
+        long savedId = clientService.saveClient(client, binService.createBin());
         if (savedId == 0){
             return new ResponseEntity<String>("Client with such login already exists", HttpStatus.OK);
         } else {
@@ -37,7 +42,7 @@ public class ClientController {
     @DeleteMapping("/{id}")
     public ResponseEntity<String> clientDelete(@PathVariable int id){
         return clientService.deleteClientFromRep(id)
-                ? new ResponseEntity<String>("Product successfully deleted", HttpStatus.NO_CONTENT)
-                : new ResponseEntity<String>("Product with id = " + id + " doesn't exists", HttpStatus.NOT_FOUND);
+                ? new ResponseEntity<String>("Client successfully deleted", HttpStatus.NO_CONTENT)
+                : new ResponseEntity<String>("Client with id = " + id + " doesn't exists", HttpStatus.NOT_FOUND);
     }
 }
