+++ Rest/src/main/java/sber/spring/Rest/service/ClientService.java
@@ -16,14 +16,10 @@ import java.util.Optional;
 public class ClientService {
     @Autowired
     private ClientRepository clientRepository;
-    @Autowired
-    private ProductRepository productRepository;
-    @Autowired
-    private BinRepository binRepository;
 
-    public int saveClient(Client client) {
+    public int saveClient(Client client, int binId) {
         if (!clientRepository.isClient(client)) {
-            return clientRepository.signClient(client, binRepository.createBin());
+            return clientRepository.signClient(client, binId);
         } else {
             return 0;
         }
@@ -37,56 +33,7 @@ public class ClientService {
         return clientRepository.deleteClient(id);
     }
 
-    public long saveProduct(Product product) {
-        return productRepository.addProduct(product);
-    }
-    public void updateProduct(Product product) {
-        productRepository.update(product);
-    }
-
-    public Optional<Product> searchProductRep(int id) {
-        return productRepository.search(id);
-    }
-
-    public boolean deleteProductFromRep(int id) {
-        return productRepository.delete(id);
-    }
-
-    public List<Product> searchProductByNameRep(String name) {
-        return productRepository.searchByName(name);
-    }
 
 
-
-    public void sellProduct(int id, int quantity) {
-        productRepository.sell(id, quantity);
-    }
-
-    public Optional<Bin> searchBinRep(int binId) {
-        return binRepository.search(binId);
-    }
-
-    public boolean isInBin(int binId, int productId) {
-        return binRepository.isInBin(binId, productId);
-    }
-
-    public int addToBin(int binId, int productId) {
-        return binRepository.add(binId, productId);
-    }
-
-    public void changeQuantity(int binId, int productId, int quantity) {
-        binRepository.changeQuantity(binId, productId, quantity);
-        //       clientRepository.updateBin(binId, binRepository.getBin(binId));
-    }
-
-    public boolean deleteProductFromBin(int binId, int productId) {
-         return binRepository.deleteFromBin(binId, productId);
-//        clientRepository.updateBin(binId, binRepository.getBin(binId));
-    }
-
-    public boolean pay(int binId) {
-        return binRepository.pay(binId);
-//        clientRepository.updateBin(binId, binRepository.getBin(binId));
-    }
 
 }
